Refactor basic area rendering to use separate component
